requirements.txt
python-telegram-bot>=22.6
selenium==4.18.1
python-dotenv==1.0.1
webdriver-manager==4.0.1